refactor: 게임 룰에 대한 관리를 GameRule이 하도록 변경
